[tests] update op names and list
