+++ tests/test_attention_ops.py
@@ -239,7 +239,7 @@ def create_kv_caches_with_random(
     return key_caches, value_caches
 
 
-@pytest.mark.test_reshape_and_cache
+@pytest.mark.reshape_and_cache
 @pytest.mark.parametrize("num_tokens", [42])
 @pytest.mark.parametrize("num_heads", [8])
 @pytest.mark.parametrize("head_size", [64, 80, 120, 256])
