+++ tests/test_general_reduction_ops.py
@@ -4,8 +4,10 @@ import torch
 import flag_gems
 
 from .accuracy_utils import (
+    ALL_FLOAT_DTYPES,
     ALL_INT_DTYPES,
     FLOAT_DTYPES,
+    POINTWISE_SHAPES,
     REDUCTION_SHAPES,
     REDUCTION_SMALL_SHAPES,
     SkipVersion,
@@ -74,6 +76,55 @@ def test_accuracy_all_dims(shape, dim, keepdim, dtype, kind):
         res_out = torch.all(inp, dim=dim, keepdim=keepdim)
 
     gems_assert_equal(res_out, ref_out)
+
+
+@pytest.mark.allclose
+@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
+@pytest.mark.parametrize("dtype", ALL_FLOAT_DTYPES + ALL_INT_DTYPES)
+@pytest.mark.parametrize("equal_nan", [False, True])
+@pytest.mark.parametrize("gen_nan", [0, 1, 2, 3, 4])
+def test_accuracy_allclose(shape, dtype, equal_nan, gen_nan):
+    # [gen_nan] 1: nan, 2: inf, 3: -inf, 4: inf vs -inf
+    rtol = torch.rand(1, dtype=torch.float32, device=flag_gems.device).item() * (
+        0.0001 if dtype in [torch.bfloat16, torch.float16] else 0.01
+    )
+    if dtype in ALL_FLOAT_DTYPES:
+        atol = (
+            torch.finfo(dtype).tiny
+            * torch.randint(0, 4, (1,), device=flag_gems.device).item()
+        )
+        inp1 = torch.full(shape, 1.234, dtype=dtype, device=flag_gems.device)
+        inp2 = torch.full(shape, 1.234, dtype=dtype, device=flag_gems.device)
+        if gen_nan:
+            nan_num = torch.full(
+                (1,),
+                float("nan" if gen_nan == 1 else "inf"),
+                dtype=dtype,
+                device=flag_gems.device,
+            )
+            # FIXME: Neg doesn't support double on torch_musa, so workaround temporarily.
+            inp1.view(-1)[0] = (
+                (-nan_num.cpu()).to(flag_gems.device) if gen_nan == 3 else nan_num
+            )
+            inp2.view(-1)[0] = (
+                (-nan_num.cpu()).to(flag_gems.device) if gen_nan >= 3 else nan_num
+            )
+    else:
+        atol = (
+            torch.finfo(torch.float16).eps
+            * torch.randint(0, 10, (1,), device=flag_gems.device).item()
+        )
+        inp1 = torch.randint(-1000, 1000, shape, device=flag_gems.device).to(dtype)
+        inp2 = torch.randint(-1000, 1000, shape, device=flag_gems.device).to(dtype)
+
+    ref_inp1 = to_reference(inp1, False)
+    ref_inp2 = to_reference(inp2, False)
+
+    with flag_gems.use_gems():
+        res_out = torch.allclose(inp1, inp2, rtol, atol, equal_nan=equal_nan)
+    ref_out = torch.allclose(ref_inp1, ref_inp2, rtol, atol, equal_nan=equal_nan)
+
+    assert res_out == ref_out
 
 
 @pytest.mark.any
