+++ tests/test_unary_pointwise_ops.py
@@ -11,7 +11,6 @@ from .accuracy_utils import (
     FLOAT_DTYPES,
     INT_DTYPES,
     POINTWISE_SHAPES,
-    SPECIAL_SHAPES,
     gems_assert_close,
     gems_assert_equal,
     to_reference,
@@ -204,22 +203,6 @@ def test_accuracy_gelu(shape, dtype, approximate):
     gems_assert_close(res_out, ref_out, dtype, atol=atol)
 
 
-@pytest.mark.glu
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_glu(shape, dtype):
-    res_inp = torch.randn(shape, dtype=dtype, device=flag_gems.device)
-    ref_inp = to_reference(res_inp, True)
-
-    for dim in range(len(shape)):
-        if shape[dim] % 2 != 0:
-            continue
-        ref_out = torch.nn.functional.glu(ref_inp, dim=dim)
-        with flag_gems.use_gems():
-            res_out = torch.nn.functional.glu(res_inp, dim=dim)
-        gems_assert_close(res_out, ref_out, dtype)
-
-
 @pytest.mark.gelu
 @pytest.mark.parametrize("shape", POINTWISE_SHAPES)
 @pytest.mark.parametrize("dtype", FLOAT_DTYPES)
@@ -255,6 +238,22 @@ def test_accuracy_gelu_(shape, dtype, approximate):
         res_out = torch.ops.aten.gelu_.default(res_inp, approximate=approximate)
 
     gems_assert_close(res_out, ref_out, dtype)
+
+
+@pytest.mark.glu
+@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_glu(shape, dtype):
+    res_inp = torch.randn(shape, dtype=dtype, device=flag_gems.device)
+    ref_inp = to_reference(res_inp, True)
+
+    for dim in range(len(shape)):
+        if shape[dim] % 2 != 0:
+            continue
+        ref_out = torch.nn.functional.glu(ref_inp, dim=dim)
+        with flag_gems.use_gems():
+            res_out = torch.nn.functional.glu(res_inp, dim=dim)
+        gems_assert_close(res_out, ref_out, dtype)
 
 
 @pytest.mark.isinf
@@ -728,62 +727,6 @@ def test_accuracy_flip_with_non_dense_input(shape, dtype, dims):
     gems_assert_equal(res_out, ref_out)
 
 
-@pytest.mark.masked_fill
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-@pytest.mark.parametrize("threshold", [0.3, 0.5, 0.7])
-@pytest.mark.parametrize(
-    "value",
-    [
-        torch.tensor(1024, device=flag_gems.device),
-        torch.scalar_tensor(1024, device=flag_gems.device),
-        1024,
-    ],
-)
-def test_accuracy_masked_fill(shape, dtype, threshold, value):
-    inp = torch.zeros(shape, dtype=dtype, device=flag_gems.device)
-    mask = torch.randn(shape, dtype=dtype, device=flag_gems.device) < threshold
-
-    ref_inp = to_reference(inp)
-    ref_mask = to_reference(mask)
-    if torch.is_tensor(value):
-        ref_out = torch.masked_fill(ref_inp, ref_mask, to_reference(value))
-    else:
-        ref_out = torch.masked_fill(ref_inp, ref_mask, value)
-    with flag_gems.use_gems():
-        res_out = torch.masked_fill(inp, mask, value)
-
-    gems_assert_equal(res_out, ref_out)
-
-
-@pytest.mark.masked_fill_
-@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-@pytest.mark.parametrize("threshold", [0.3, 0.5, 0.7])
-@pytest.mark.parametrize(
-    "value",
-    [
-        torch.tensor(1024, device=flag_gems.device),
-        torch.scalar_tensor(1024, device=flag_gems.device),
-        1024,
-    ],
-)
-def test_accuracy_masked_fill_(shape, dtype, threshold, value):
-    inp = torch.zeros(shape, dtype=dtype, device=flag_gems.device)
-    mask = torch.randn(shape, dtype=dtype, device=flag_gems.device) < threshold
-
-    ref_inp = to_reference(inp)
-    ref_mask = to_reference(mask)
-    if torch.is_tensor(value):
-        ref_inp.masked_fill_(ref_mask, to_reference(value))
-    else:
-        ref_inp.masked_fill_(ref_mask, value)
-    with flag_gems.use_gems():
-        inp.masked_fill_(mask, value)
-
-    gems_assert_equal(inp, ref_inp)
-
-
 TILE_DIMS = [(0,), (2,), (2, 0), (0, 2), (2, 2), (2, 2, 2), (2, 2, 2, 2)]
 
 
@@ -856,36 +799,7 @@ def test_accuracy_log(shape, dtype):
     gems_assert_close(res_out, ref_out, dtype)
 
 
-@pytest.mark.fill_
-@pytest.mark.parametrize("value", [0, 1, 9])
-@pytest.mark.parametrize("shape", SPECIAL_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_fill_(value, shape, dtype):
-    # Test fill_.Scalar
-    x = torch.ones(shape, device=flag_gems.device, dtype=dtype)
-    ref_x = to_reference(x.clone(), False)
-
-    ref_x.fill_(value)
-    with flag_gems.use_gems():
-        x.fill_(value)
-
-    gems_assert_equal(x, ref_x)
-
-    # Test fill_.Tensor
-    x = torch.ones(shape, device=flag_gems.device, dtype=dtype)
-    ref_x = to_reference(x.clone(), False)
-    value_tensor = torch.tensor(value, device=flag_gems.device, dtype=dtype)
-    if flag_gems.device == "musa":
-        ref_x.fill_(value_tensor.cpu())
-    else:
-        ref_x.fill_(value_tensor)
-    with flag_gems.use_gems():
-        x.fill_(value_tensor)
-
-    gems_assert_equal(x, ref_x)
-
-
-@pytest.mark.to_dtype
+@pytest.mark.to
 @pytest.mark.parametrize("shape", POINTWISE_SHAPES)
 @pytest.mark.parametrize("dtype", ALL_FLOAT_DTYPES + ALL_INT_DTYPES)
 def test_accuracy_to_dtype(shape, dtype):
